feat: adding blinker
