+++ lib/src/widgets/widgets.dart
@@ -1,4 +1,5 @@
 export 'dialogs/dialogs.dart';
+export 'nes_blinker.dart';
 export 'nes_button.dart';
 export 'nes_check_box.dart';
 export 'nes_container.dart';
